requirements.txt
streamlit
pandas
numpy
sklearn
waterfallcharts
treeinterpreter
matplotlib